Get requirements from stack model
